+++ src/main/java/com/group3/wineshop/entities/Type.java
@@ -1,8 +1,6 @@
 package com.group3.wineshop.entities;
 
 import javax.persistence.*;
-import java.util.LinkedHashSet;
-import java.util.Set;
 
 @Entity
 @Table(name ="type")
@@ -10,7 +8,7 @@ import java.util.Set;
 public class Type {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
     private String name;
 
     public Type() {
@@ -20,11 +18,11 @@ public class Type {
     this.name = name;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
