+++ src/main/java/com/group3/wineshop/repositories/TypeRepository.java
@@ -4,5 +4,5 @@ import com.group3.wineshop.entities.Type;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface TypeRepository extends JpaRepository<Type, Long> {
+public interface TypeRepository extends JpaRepository<Type, Integer> {
 }
