+++ src/main/java/com/group3/wineshop/repositories/WineRepository.java
@@ -3,6 +3,6 @@ package com.group3.wineshop.repositories;
 import com.group3.wineshop.entities.Wine;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WineRepository extends JpaRepository<Wine, Long> {
+public interface WineRepository extends JpaRepository<Wine, Integer> {
 
 }
